Overhaul of README
Major update to the README:

Now reflects when it was last updated and has two additional sections for functionality that is not currently working as intended but is implemented and functionality that is not yet implemented but desired. Additionally, various definitions updated to reflect updates made to commands and to change terminology for clarity.